% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pattern.R
\name{find_pattern}
\alias{find_pattern}
\alias{is_in_file}
\title{Find a pattern in files from a directory}
\usage{
find_pattern(pattern, where = here(), full_names = FALSE)

is_in_file(pattern, file)
}
\arguments{
\item{pattern}{the pattern to find. Can be regex}

\item{where}{the path to the directory where you want to search}

\item{full_names}{a logical value as in \code{\link[base]{list.files}}. If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned.}

\item{file}{the path to the file where you want to search}
}
\value{
a vector with all the files where the pattern was found for \code{find_pattern} or a logical value for \code{is_on_file}
}
\description{
\code{find_pattern} search a pattern in all files from a directory and \code{is_on_file} search
for a pattern in one file.
}
\examples{
find_pattern(pattern = "usethis::", where = system.file(package = "benutils"))

\dontrun{
# if you are in a R project you can just specify the pattern
find_pattern("my_pattern")
}
}
