% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_documentation.R
\name{make_data_documentation}
\alias{make_data_documentation}
\alias{document_data}
\title{Make template documentation for each internal dataset durong a package development}
\usage{
make_data_documentation(where = here())

document_data(file)
}
\arguments{
\item{where}{path to the directory where tha package is developped}

\item{file}{path to the file (.rda) of the dataset to document}
}
\description{
BEWARE : this will overwrite existing documentation (I need to work on that)
}
\examples{
\dontrun{
# In a R project where you develop the package
make_data_documentation()
}
}
